% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh.com-package.R
\docType{package}
\name{gh.com-package}
\alias{gh.com}
\alias{gh.com-package}
\title{gh.com: scrape github}
\description{
gh.com makes it easy to search for and retrieve information about
repositories github.com
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mkearney/gh.com}
  \item Report bugs at \url{https://github.com/mkearney/gh.com/issues}
}

}
\author{
\strong{Maintainer}: Michael W. Kearney \email{kearneymw@missouri.edu} (0000-0002-0730-4694)

}
\keyword{internal}
