% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape.R
\name{scrape_github}
\alias{scrape_github}
\title{Scrape github.com}
\usage{
scrape_github(query, pages = NULL)
}
\arguments{
\item{query}{Search query to perform on github.com}

\item{pages}{Max number of pages, defaults to all.}
}
\value{
Search results in the form of a tibble data frame where each row
  corresponds to a search result and the columns contain information about
  the matching repo (id, name, url, description, language, stars, license,
  topics
}
\description{
Returns tibble of search results
}
