% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape.R
\name{scrape_rstudio_themes}
\alias{scrape_rstudio_themes}
\title{Scrape rstudio theme repos on github.com}
\usage{
scrape_rstudio_themes()
}
\value{
A tibble of results from search designed to find rstudio themes
}
\description{
Returns tibble of rstudio theme search results
}
