% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh.com-package.R
\docType{package}
\name{gh.com-package}
\alias{gh.com}
\alias{gh.com-package}
\title{gh.com: scrape github}
\description{
gh.com makes it easy to search for and retrieve information about
repositories github.com
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (YOUR-ORCID-ID)

}
\keyword{internal}
